`MakeFixedFlatMap` now `consteval`
By making this consteval, the issue of sorting has been sidestepped with
a more reliable approach. The use of `base::sorted_unique` is not
required anymore for fixed flat maps, as consteval guarantees the
operation will happen at compile time, as opposed to the compiler
bailing out during sorting with `constexpr`.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/45e5a3254798f4f5ccb1ee5b220513a6b773f529

commit 45e5a3254798f4f5ccb1ee5b220513a6b773f529
Author: Andrew Grieve <agrieve@chromium.org>
Date:   Tue Jan 2 18:36:08 2024 +0000

    Make base::MakeFixedFlatMap consteval

    This ensures that it will be evaluated at compile-time.

    Also adds MakeFixedFlatMapNonConstEval for the one instance where
    compile-time evaluation is not possible.

    Bug: None