Fix CommanderServiceBrowserTest
There was a refactoring of the AutocompleteController which caused some tests to fail. There were two causes:

A DCHECK that we weren't NotifyListeners wasn't called if the Provider was done in the controller.
OmniboxView->SetText isn't immediately reflected in GetText, so there is a small refactor for making the Provider pass the current text.
There's a weird going on where the ... has a space before it. I think that's fine to address separately, it doesn't look like the string has changed.