[Android] New c-tor signature for TabbedRootUiCoordinator
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/d3ca60c16d6cf77538ea7b3e8b19cef537cf3bef

[Hub][Tab Switcher] Fix crash on back press of incognito reauth

Invoking LayoutManager show when the Hub is already visible (via
TAB_SWITCHER layout type crashes or shows visual artifacts).

Instead the IncognitoReauth backpress handler should return to the
normal tab switcher pane if the "Hub" is visible.

As additional context. Hub is a new "container" for the TabSwitcher it
already supports IncognitoReauth with the same APIs as TabSwitcher
this is just a fix for switching between Incognito and Regular modes
when in the Hub.

Bug: b/321016874