[Android] New pref in safety check settings
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/004b0920380576bd04e12e8f4a37d5af106bc158

[UPMLocalSettings] Introduce the UI for the second password check

This CL introduces a new password check preference into
safety_check_preferences.xml. When there will be split stores, one
preference will be used to display the result of the password check in
the local password storage, another - for the account password storage.
SafetyCheckSettings fragment also has the logic to set visibility of the
local and account tied preferences depending on whether the appropriate
password storage is available.

Bug: b/306670833